Adjust attn output dtype in all architectures

+++ exllamav3/architecture/ernie4_5.py
@@ -89,7 +89,8 @@ class Ernie4_5Model(Model):
                     key_k = "k_proj",
                     key_v = "v_proj",
                     key_o = "o_proj",
-                    qmap = "block.attn"
+                    qmap = "block.attn",
+                    out_dtype = torch.float,
                 ),
                 mlp_norm = RMSNorm(
                     config = config,
